+++ src/LiteGraph/Client/Implementations/NodeMethods.cs
@@ -452,9 +452,9 @@
 
         #endregion
 
-        #region Private-Methods
+        #region Internal-Methods
 
-        private Node PopulateNode(Node obj, bool includeSubordinates, bool includeData)
+        internal Node PopulateNode(Node obj, bool includeSubordinates, bool includeData)
         {
             if (obj == null) return null;
 
@@ -472,6 +472,10 @@
             if (!includeData) obj.Data = null;
             return obj;
         }
+
+        #endregion
+
+        #region Private-Methods
 
         #endregion
     }